configmgr: add SAL_CONFIG_WINREG_RETAIN_TMP variable.
Windows registry dumping behaves very strangely, and in hard to debug
ways - this should help in-the-field debugging of this in future.

+++ configmgr/source/components.cxx
@@ -561,7 +561,8 @@ Components::Components(
             OUString aTempFileURL;
             if (dumpWindowsRegistry(&aTempFileURL, eType)) {
                 parseFileLeniently(&parseXcuFile, aTempFileURL, layer, nullptr, nullptr, nullptr);
-                osl::File::remove(aTempFileURL);
+                if (!getenv("SAL_CONFIG_WINREG_RETAIN_TMP"))
+                    osl::File::remove(aTempFileURL);
             }
             ++layer; //TODO: overflow
 #endif
